Stage the merge conflict fix that a prior commit missed
internal/search/client_test.go's deletion and the new testhelpers_test.go
(the still-needed fakeSearchIndex helper, extracted before deleting the
file) were made while resolving the merge but never git-add'ed before
the merge commit — CI caught it (same "unknown field semantic" vet
failure the merge itself was supposed to fix). Verified locally this
time: build/vet/vet-integration clean.
